[rename/#47] pass_card -> passcard
